feat: move squad created action to squad created worker
