+++ src/workers/sourceSquadCreated.ts
@@ -3,8 +3,9 @@ import {
   addUserToContacts,
   LIST_DRIP_CAMPAIGN,
 } from '../common';
-import { Source, SourceMember, User } from '../entity';
+import { Source, SourceMember, User, UserActionType } from '../entity';
 import { SourceMemberRoles } from '../roles';
+import { insertOrIgnoreAction } from '../schema/actions';
 import { ChangeObject } from '../types';
 import { messageToJson, Worker } from './worker';
 
@@ -13,7 +14,7 @@ interface Data {
 }
 
 const worker: Worker = {
-  subscription: 'api.source-squad-created-mailing',
+  subscription: 'api.source-squad-created',
   handler: async (message, con, logger) => {
     const data: Data = messageToJson(message);
 
@@ -40,6 +41,8 @@ const worker: Worker = {
       if (!user) {
         return;
       }
+
+      await insertOrIgnoreAction(con, user.id, UserActionType.CreateSquad);
 
       if (user.acceptedMarketing) {
         const contactId = await getContactIdByEmail(user.email);
